Reorganize files in folders
