+++ src/bpmn/end_event.js
