+++ src/core/Element.js
